agente_entrevistas/requirements.txt
# ── Google ADK + Claude (via LiteLLM) ────────────────────────────────────────
google-adk>=0.3.0
anthropic>=0.40.0
litellm>=1.40.0

# ── Supabase ──────────────────────────────────────────────────────────────────
supabase>=2.4.0

# ── Email ───────────────────────────────────────────────
mailtrap>=2.0.0

# ── Búsqueda web ──────────────────────────────────────────────────────────────
tavily-python>=0.3.3

# ── Generación de documentos ──────────────────────────────────────────────────
python-docx>=1.1.0
# El .pdf se convierte con LibreOffice (opcional)

# ── Validación ────────────────────────────────────────────────────────────────
pydantic>=2.6.0

# ── Utilidades ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
PyMuPDF>=1.24.0

# ── Observabilidad (opcional) ─────────────────────────────────────────────────
langfuse>=2.30.0

# ── Tests ─────────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
